Move server actions file to the actions file.
